Clean and split hdf5 subpackage
  - Remove postqe related functions
  - Keep HDF5 utils in hdf5.py
  - Keep UPF utils in upf.py
  - Delete old hdf5/ subpackage
  - Add h5py to package requirements
