--- qeschema/hdf5/__init__.py
@@ -1,18 +1,0 @@
-#
-# Copyright (c), 2021, Quantum Espresso Foundation and SISSA (Scuola
-# Internazionale Superiore di Studi Avanzati). All rights reserved.
-# This file is distributed under the terms of the MIT License. See the
-# file 'LICENSE' in the root directory of the present distribution, or
-# http://opensource.org/licenses/MIT.
-#
-from .charge import read_charge_file_hdf5, get_minus_indexes, \
-    get_charge_r, write_charge
-from .readutils import get_wf_attributes, get_wavefunctions, \
-    get_wfc_miller_indices, read_pseudo_file, create_header, \
-    read_postqe_output_file, read_etotv
-
-__all__ = [
-    'read_charge_file_hdf5', 'get_minus_indexes', 'get_charge_r', 'write_charge',
-    'get_wf_attributes', 'get_wavefunctions', 'get_wfc_miller_indices',
-    'read_pseudo_file', 'create_header', 'read_postqe_output_file', 'read_etotv'
-]
