+++ setup.py
@@ -16,7 +16,7 @@ with open("README.rst") as readme:
 setup(
     name='qeschema',
     version='1.3.0',
-    install_requires=['xmlschema>=1.6.4', 'pyyaml', 'numpy'],
+    install_requires=['xmlschema>=1.6.4', 'pyyaml', 'numpy', 'h5py'],
     packages=['qeschema'],
     package_data={'qeschema': ['schemas/*.xsd', 'schemas/releases/*.xsd']},
     scripts = ['scripts/xml2qeinput.py', 'scripts/yaml2qeinput.py'],
